chore: fix directory structure
