+++ src/main/java/de/uos/informatik/ko/rcp/Main.java
@@ -11,39 +11,50 @@ import de.uos.informatik.ko.rcp.generators.SerialScheduleGenerator;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Random;
 
 public class Main {
-    public static void main(String[] args) {
+
+    public static void main(String[] args){
         if (args.length != 1) {
-            System.err.println("Usage: solver <instance path>");
-            System.exit(1);
+            System.out.println("Wrong usage");
+            return;
         }
 
-        var instance = Io.readInstance(Paths.get(args[0]));
+        final String path = args[0];
 
-        final var essGenerator = new EarliestStartScheduleGenerator(instance);
+        final Instance instance = Io.readInstance(Paths.get(path));
 
-        int[] order = new int[instance.n()];
-        for (int i = 0; i < instance.n(); ++i) {
-            order[i] = i + 1;
+        EarliestStartScheduleGenerator EssGen = new EarliestStartScheduleGenerator(instance);
+
+
+        for (int i = 0; i < instance.r(); i++) {
+            System.out.printf("res %d has %d available units\n", i, instance.resources[i]);
         }
 
-        final var start = System.currentTimeMillis();
-        final var schedule = essGenerator.generateSchedule(order);
-        final var end = System.currentTimeMillis();
+        //TODO richtigen Random vom Nutzer nutzen
+        Random random = new Random();
+        int[] solution = new int[instance.n()];
 
-        System.out.println("Schedule:");
-        for (int actIdx = 1; actIdx < instance.n() - 1; ++actIdx) {
-            System.out.println(
-                    "Activity " + actIdx + ": start time = " + schedule[actIdx]
-                    + ", processing time = " + instance.processingTime[actIdx]);
+        System.out.println();
+
+        int[] reihenfolge = new int[instance.n()];
+
+        // TODO Deep Copy nötig? Eigentlich nicht, oder?
+        reihenfolge = GA.geneticAlgorithm(instance, random);
+
+        for (int i = 0; i < reihenfolge.length; i++) {
+            System.out.println(reihenfolge[i]);
         }
 
-        System.out.println("Makespan: " + schedule[instance.n() - 1]);
+        solution = EssGen.generateSchedule(reihenfolge);
 
-        final boolean admissible = Utils.checkAdmissibility(instance, schedule);
-        System.out.println("Admissible? " + admissible);
+        System.out.println("Makespan" + solution[solution.length-1]);
 
-        System.out.println("Took " + (end - start) + " milliseconds.");
+        Io.writeSolution(solution, Paths.get("sol2.txt"));
+
+        System.out.println("ok");
+
     }
+
 }
